# internal/server/docs/swagger.yaml
components:
  schemas:
...
    api.APIError:
      properties:
        code:
          type: string
        details:
          additionalProperties: {}
          type: object
        message:
          type: string
      type: object
...
    api.ErrorEnvelope:
      properties:
        error:
          $ref: '#/components/schemas/api.APIError'
      type: object
...
    api.ModifyResult:
      properties:
        changeId:
          type: string
        recordIds:
          items:
            type: string
          type: array
          uniqueItems: false
        rejections:
          items:
            $ref: '#/components/schemas/api.OpRejection'
          type: array
          uniqueItems: false
        versionId:
          type: string
      type: object
...
    api.OpRejection:
      properties:
        opIndex:
          type: integer
        reason:
          type: string
        recordId:
          type: string
        recordIndex:
          type: integer
      type: object
...
    api.PropertiesGetResponse:
      properties:
        record:
          additionalProperties: {}
          type: object
      type: object
...
    api.PropertiesSetRequest:
      properties:
        patch:
          additionalProperties: {}
          type: object
      type: object
...
info:
  description: Local HTTP/JSON API wrapping any-sync-sdk. Localhost-only, no auth
    in v1.
  title: Any API
  version: "1.0"
openapi: 3.1.0
paths:
...
  /spaces/{spaceId}/properties/{objectId}:
    get:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Object ID
        in: path
        name: objectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.PropertiesGetResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Get properties of an object
      tags:
      - properties
  /spaces/{spaceId}/properties/{objectId}/collections/{collectionId}:
    delete:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Object ID
        in: path
        name: objectId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ModifyResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Remove the object from a collection
      tags:
      - properties
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Object ID
        in: path
        name: objectId
        required: true
        schema:
          type: string
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ModifyResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Add the object to a collection
      tags:
      - properties
  /spaces/{spaceId}/properties/{objectId}/set/{typeId}:
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Object ID
        in: path
        name: objectId
        required: true
        schema:
          type: string
      - description: Type or collection ID (the owner)
        in: path
        name: typeId
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.PropertiesSetRequest'
                description: Patch map
                summary: body
        description: Patch map
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ModifyResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Set properties on an object (single declared scope)
      tags:
      - properties
  /spaces/{spaceId}/properties/{objectId}/type:
    delete:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Object ID
        in: path
        name: objectId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ModifyResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Unset the object's type
      tags:
      - properties
  /spaces/{spaceId}/properties/{objectId}/type/{typeId}:
    post:
      parameters:
      - description: Space ID
        in: path
        name: spaceId
        required: true
        schema:
          type: string
      - description: Object ID
        in: path
        name: objectId
        required: true
        schema:
          type: string
      - description: Type ID
        in: path
        name: typeId
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ModifyResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.ErrorEnvelope'
          description: Internal Server Error
      summary: Set the object's type
      tags:
      - properties
...
servers:
- url: http://127.0.0.1:7001/v1